Move API layer code into `api` module
