+++ backend/src/api.rs
@@ -1,0 +1,6 @@
+pub mod error;
+pub mod dto;
+pub mod handler;
+pub mod middleware;
+pub mod router;
+pub mod validated_json;
