+++ backend/src/api/dto/responses.rs
@@ -1,3 +1,4 @@
+use crate::models::post::PostWithAuthor;
 use serde::{Deserialize, Serialize};
 
 /// A general-purpose error response struct.
@@ -34,4 +35,21 @@ pub struct PostResponse {
     pub edited_at_ms: Option<i64>,
     pub archived_at_ms: Option<i64>,
     pub deleted_at_ms: Option<i64>,
+}
+
+impl From<PostWithAuthor> for PostResponse {
+    fn from(pwa: PostWithAuthor) -> Self {
+        Self {
+            id: pwa.id,
+            author_username: pwa
+                .author_username
+                .unwrap_or_else(|| String::from("[deleted]")),
+            parent_id: pwa.parent_id,
+            body: pwa.body.unwrap_or_else(|| String::from("[deleted]")),
+            created_at_ms: pwa.created_at.timestamp_millis(),
+            edited_at_ms: pwa.edited_at.map(|ms| ms.timestamp_millis()),
+            archived_at_ms: pwa.archived_at.map(|ms| ms.timestamp_millis()),
+            deleted_at_ms: pwa.deleted_at.map(|ms| ms.timestamp_millis()),
+        }
+    }
 }
